mod2222

+++ capstone/tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountDao.java
@@ -15,5 +15,5 @@ public interface AccountDao {
 
 //    Account get(BigDecimal balance);
 
-    BigDecimal getBalance(BigDecimal balance);
+    int getBalance(int balance);
 }
